Use DynamicZip in group by to allow group-by by more than 5 fields.
WARNING: In order to achieve this the signature for `hash_join::prepare_hashed_relation`
and `group_by::groupby` have been modified and they do not constraint the type to be
copyable anymore. Then, move can occur when using this function. This change is needed as
DynamicZip returns a Vec and not a tuple as Zip do.