improve comparison in multiple key hashing
Groupby and join algorithm uses a custom hash when multiple keys are hashed.
Instead of the values, the idx of the rows are stored in the hashtable as keys.
These indexes are used to for comparison during probing/ insertion of the
hash tables. This commit improves that probing functions performance by ~20%.