fork rayon par_sort and improve argsort perf
This argsort implementation also keeps track of the indexes
during the sort procedure. This remains cache efficient because
compared/swapped elements are always close. It still needs allocation
of double T and Idx buffers, but it saves one fine allocation that
was needed to unzip (T, Idx) as the two are now already separated.

Also a few improvements in default sort and vec creation (e.g. memcpy).