Fix test_cum_agg + add example Series.cumprod
The current version of `test_cum_agg`, operating on the series `[1, 2]`, compares the output of the `cum*` functions to a list using `==`, which results in a new series which may have both True's and False's; assert gives it a pass. This is most visible with the original asserted output of cumsum to be `[1, 2]` whilst it should be `[1, 3]` obviously.

This PR:
* uses `.series_equal` to prevent the wrong comparison
* adds an assertion on `cumprod`, which was not covered yet
* adds an example to the docstring of cumprod
* makes the test case slightly more interesting by using a 4 item series as input