Add `u64` as index type (#2682)
This allows polars to compile with a feature gate bigidx, which replaces the current limit of u32::MAX rows to u64::MAX rows in polars.

Now I am curious how long it will take before someone needs to compile polars this way. ;)

Note that the default index is faster and consumes less memory, so don't activate it unless you really need it.