reduce generic bloat
A lot of `take_*` methods are called with
an iterator over &[IdxSize] take simply
applies a closure `Fn(v: &IdxSize) -> *v as usize`.
Assuming that every closure is a duplication of the
take_* functions this is a lot of bloat.

Now we implement the conversion to `TakeIdx` that uses
a function item, so one instantiation.