Literal in groupby context, arange and repeat
1. This fixes the arange function to work in
the groupby context.

2. We improve performance and memory usage of literals in grouping
By doing so we found an inconstency in the literal expansion which
lead to the following expression working in grouping:

lit(1).cumsum() would lead to a list of the group length.

This now does what you expect: e.g. [1].

To get old behavior we introduce a new expression: `repeat`

repeat(1, count()).cumsum()