Update cummax documentation
Right now, the cummax() documentation uses [1, 2, 3] as its example series. This series isn't the best short example because its cummax is also [1, 2, 3]. It's not immediately clear why passing that as input gets you identical output. The reason the output is identical to the input is that [1, 2, 3] happens to be a monotonically increasing sequence.

The slight change I'm introducing here is an example Series that's *not* monotonically increasing. I think this choice might help users better understand what cummax actually does.