Assign dtypes to expected columns when dtypes is a list and column selection is used when reading a CSV file.
Assign dtypes to expected columns when dtypes is a list and column
selection is used when reading a CSV file.

As dtypes are passed to the polars CSV reader, projection and column
selection was not applied on the dtypes in case it is a list,
so the dtypes would be applied to the first x columns, instead of
the corresponding columns selected by the projection/column selection.

A remaining issue that is not fixed:
  Projection in combination with a list of dtypes which contain a
  pl.Date type, are converted to null dates.
  In case column names were used for the selection, the pl.Date type
  will correctly convert that column to dates.

Fixes: #3891