fix[python]: Fix type indexing py3.7 test error
In #4058, I enforce strict type equality checks in mypy. Most cases can be fixed with `typing.cast`. For generic types such as tuple, we cannot use this yet, because py3.7 does not allow this, and thus we need a `type: ignore`. This one case of using the cast slipped through, fixing this here.

As noted by https://github.com/pola-rs/polars/pull/4413#issuecomment-1214400239