feat[python]: Improve numpy ufunc dtype support in expressions (#3228, #5012)
As numpy ufunc dtype detection for series was improved/fixed in #3583,
direct dispatch can be done from the numpy ufunc implementation
on an expression to the numpy ufunc implementation on a series,
without the need to infer the dtype there.

Overriding the output dtype for numpy ufunc on series was not working,
and is fixed too.