fix(python): Check multiprocessing start_method on import (#3144)
Check multiprocessing start_method on import of Polars and set
python multiprocessing start method to "spawn" if undefined.
or raise ImportError if defined, but not set to "spawn".

Python multiprocessing with start method "fork" can not be used
as forking does not copy threads (Polars spawns multiple threads
for its rayon threadpool, resulting in locks that can't
be released, resulting in a hanging python interpreter.