feat(python): Add .item() on DataFrame and Series
Requested on Discord to alleviate the syntax troubles when a dataframe or series is just one element, and you end up writing `df[0,0]` or `s[0]` once too many.

.item() is taken from numpy: https://numpy.org/doc/stable/reference/generated/numpy.ndarray.item.html
I have on purpose not allowed an input argument. We have our indexing machinery and syntax, and this method is purely for converting to a scalar. Otherwise people start to write `s.item(3)` rather than `s[3]`. Having said that, maybe we should therefore call this method `to_scalar`, that is probably more clear, including that it will error if it cannot be convert to a scalar. Although if the single entry would be a string, is "scalar" the right terminology? Open to suggestions.