fix(rust, python): remove uses of rayon global thread pool
Found a few spots where the rayon global thread pool is being used, and swapped them out with the internal pool. I was able to track down the spots by using a local version of rayon and adding a panic statement in the `set_global_registry` method which is invoked whenever the global thread pool is constructed. I'm not sure what the best way to enforce this is. One idea is to use test fixtures, and we can check if the global pool was used at the end of each test. Another could be to patch rayon when testing similar to what I did to detect the errors.