Detect deltalake version in show_versions
I noticed that we did not pick up the version of the optional deltalake dependency, as per https://github.com/delta-io/delta/issues/1339.

Solved this using importlib.metadata. This standard library function is not availble for Python 3.7, but I dont think it is worth adding the backport for that given support may soon be lost and there is no significant loss in functionality to end users.