feat(python): Allow construction of Series with list of only None with specific dtype via dtype_if_empty.
Allow construction of Series with list of only `None` with specific dtype via
`dtype_if_empty` like it currently is supported for just an empty list constructor,
so the default dtype is not pl.Float64 anymore.

    pl.Series([None, None, None], dtype_if_empty=pl.Int8).dtype == pl.Int8

This is feature is now indirectly used in `map_dict`, so the following,
now returns a more suited dtype than pl.Float64.

     pl.Series("a", [42], dtype=pl.Int8).map_dict({}, default=pl.first()).dtype == pl.Int8
     pl.Series("a", [42], dtype=pl.Int8).map_dict({5: None}, default=pl.first()).dtype == pl.Int8

Closes: #7705