feat(python): Add mean, median and quantile to Expr.dt
Currently, Expr.min() and Expr.max(), for date and datetime dtypes, return the minimum and maximum. However, for mean, median and quantile, null is returned. This PR adds these methods in DateTime namespace, similar to what has been done for `Series`.

One problem I run into is that the cast to physical type is easy, but then on the way out how to figure out what the dtype should be? Right now I have hard-coded `Date`. Is there a way to get the dtype from an expression? @ritchie46