chore,doc(python): Address ignored Ruff doc rules
D101, D102 & D103 are all useful rules, so have added docstrings where possible. On some internal utils, I have marked private and/or used a noqa. This should stop us from adding undocumented public functions.

D100 & D104: we are not using module and package docstrings at all, is this something we want/need? As it is in the list of "remove errors" in pyproject.toml.

D105: we should probably do this, but there are a lot here.