fix(rust) fix build for wasm
Fixes issue #8783, where v0.29 broke the build for wasm.
In addition to fixing the build, this adds a CI test to prevent
regressions like this in the future.

Relatedly there's an issue for the `home` crate rust-lang/cargo#12297 to
make it compile for wasm. The previously used crate there, `dirs`,
_does_ support wasm (just returning None for the home dir).