Improved ErrorState::take() logic
Instead of leaving a `String` in place of `PolarsError`, leave in place
the original `PolarsError`. Then, when it's time to produce a new error,
we clone the original error if possible; and it's almost always
possible, as all `PolarsError` variants contain ErrString as their data
except one: `Io`. So just convert `Io` to `ComputeError` with an error
message saying so, otherwise reuse the `PolarsError` variant with a
cloned message. (This is based off, and utilizes, the method
`PolarsError::wrap_msg`.)

This way we can retain the "shape" of the original error and just
augment its message to indicate that it's already failed n times.

Also cleaned up the `err_msg.rs` test in anticipation of adding a bit
more to that file.