Simplified implementation of error reporting
Now that errors aren't taken and replaced by a String - the original is
stored and cloned as needed - we can simplify the implementation.
Instead of NotYetEncountered and AlreadyEncountered, we can just have a
single struct with a counteer (no enum needed). This massively
simplifies the implementation
Also, simplified the implemtnation of tests
Also cleaned up ErrorState's Debug implementation so that users don't
see the nuts and bolts inside the struct - they only see the important
debugging info