docs(python): clarify behavior of `DataFrame.rows_by_key`.
When I read the docs for this method for the first time, it was not at
all clear to me how the returned dictionary was constructed until I
scrolled all the way down to the examples. And even then, it took me
a little while to understand.

I edited the description to be a bit more clear about how the dictionary
is structured, while trying not to be too redundant with the examples and
parameter docs.

Also, while I was here, I added a little blurb to the `rows` docs as well.