perf: prune parquet row groups when `is_not_null` is used
Filtering on non null values is more used in the wild than filtering on null values.

I changed the `BatchStats` to include the number of rows in that rg and compare it with
the number of nulls statistic if avaialable.

I also tried to simplify `not(is_null)` to `is_not_null` as I see this pattern a lot,
this will allow to apply the pruning in these cases.