Refactor CSV serialization to not go thorough `AnyValue`
While convenient, this harms performance. Avoiding it gives a boost of 40% on my machine, and even more if I limit to single thread.

I did several other things to improve performance, like pre-compilation of date/time formatting and speeding up strings with quotes that need to be escaped. Those aren't strictly related, but they are nice.