feat(rust): handle CSE dtypes in NodeTraverser.get_dtype
Since the CSE expressions don't appear in the schema that IR::Select
and IR::HStack publically advertise, deducing the dtype of expressions
that depend upon them fails. To handle this, augment the public schema
with the dtypes of the CSE terms before asking for the dtype of the
remaining expressions.