fix(rust): small safety issue in CWC filtermap
I woke up in the middle of the night and realized I wrote an unsound implementation here. This fixes an issue in CWC where some unsafe code could potentially lead to unsoundness. This should never actually have triggered. But it the spirit of keeping things maintainable in the future, I thought I would properly fix it.