feat(rust): add `env` locked metadata functions
This prepares for further use of metadata. This locks several functionalities
of the metadata behind the `POLARS_METADATA_FLAGS` environment variable when
compiling on `debug`-mode. This is either `0` to disable all metadata, or
`extensive` and `log` possibility delimited by `,`. This enables further
experimental features like `extensive` use of the metadata and `log`ging of
every time the metadata is used.