fix(rust): fix parquet read/write nested
This PR rewrites many of the parts that relate to Parquet reading and writing of nested fields. This PR is quite large, and I tried my best to extensively test the new implementation, but I am also confident that there are still bugs in it. My view at the moment that these bugs are fringe enough for the moment.

Specifically, this PR merges the `RepLevelIter` and `DefLevelIter` into a `BufferedDremelIter` that properly follows the Parquet specification when it comes to *repetition* and *definition* levels. It is also buffered and thus should be faster (although I don't have numbers to back this up). For the reader side, most of the code is adjusted to skip add elements that are implied by invalid higher-level nestings.

This PR also does many, many drive-by refactors in the Parquet code.

Fixes #16608.