feat: allow experimental metadata use on release
This adds changes the metadata flag from `POLARS_METADATA_FLAGS` to
`POLARS_METADATA_USE`. I also renamed `extensive` to `experimental`. There are
3 settings that can be controlled.

| Setting      | Debug Build                        | Release Build                      |
|--------------|------------------------------------|------------------------------------|
| Enabled      | `POLARS_METADATA_USE=1` (default)  | `true`                             |
| Experimental | `POLARS_METADATA_USE=experimental` | `POLARS_METADATA_USE=experimental` |
| Logging      | `POLARS_METADATA_USE=log`          | `false`                            |

In the debug-mode, the value can be changed during runs to make A/B testing
possible. For release-mode builds, the value is cached and cannot be changed
after it is set initially.

There are 5 accepted values.

* `0`: Turn off all usage of metadata
* `1`: Turn on usage of metadata (default)
* `experimental`: Turn on normal and experimental usage of metadata
* `experimental,log`: Turn on normal, experimental usage and logging of metadata usage
* `log`: Turn on normal and logging of metadata usage

In general, the runtime overhead should be negligible since the value is cached
and branch prediction should do it work on here.