fix: validate column names in drop by default
This PR adds the `strict` flag on `LazyFrame::drop` that is set to `True` by default. When it is enabled, `drop` will validate the given column names against the schema and throw a `ColumnNotFound` error if the column is not defined instead of quietly ignore it. The flag can be set to `False` to re-enable the previous quiet ignoring behavior.