feat: predicate + projection pushdown in NDJson
This PR adds predicate and projection pushdown in NDJson in a similar fashion as is done in CSV. This currently just filters after each chunk is deserialized, which is not perfectly efficient. It would probably be a good idea to do the projection during the `into_series` call and the predicate during the `...ChunkedBuilder` append value.