fix(rust): bubble error when no available bitrepr
This PR refactors the code surrounding bit representations and properly bubbles
errors up if no bit representation is defined.

This resolves the panic in #14826, but does not implement the wanted behavior.

I am not sure a test case is useful here as the behavior should be added in the end.