perf(rust): faster decode on Parquet HybridRLE
This PR performs several optimization surrounding the HybridRLE decoding:

- A lot of iterators in favor of collecting and buffering
- Flatten nested iterators into a single iterators
- Remove many memcopies in `bitpacked::decode`.

Overall, my microbenchmarks showed an improvement of roughly 10x for the HybridRLE decoding. In order to get a better view on how to much impact this has on totality of parquet decoding I did a benchmark using the Yellow NYC Taxi dataset. This got me a roughly 1.8x speed-up.