perf: batch parquet primitive decoding
This is rather large change that changes quite fundamentally how HybridRLE and
Parquet decoding works. There are now two important concepts that speed up the
performance of the Parquet reader while utilizing less memory than before. This
does how increase the complexity of the code.

First, for a benchmark using the NYC Yellow-Taxi dataset (we decode the whole
dataset 100x). Here, we see the following results.

No maximum threads:

```
Benchmark 1: After Optimization
  Time (mean ± σ):      4.918 s ±  0.076 s    [User: 28.748 s, System: 2.486 s]
  Range (min … max):    4.819 s …  5.064 s    10 runs

Benchmark 2: Before Optimization
  Time (mean ± σ):      7.333 s ±  2.144 s    [User: 60.374 s, System: 3.054 s]
  Range (min … max):    5.416 s … 11.132 s    10 runs

Summary
  After Optimization ran
    1.49 ± 0.44 times faster than Before Optimization
```

Maximum threads = 1:

```
Benchmark 1: After Optimization
  Time (mean ± σ):     18.452 s ±  0.054 s    [User: 16.058 s, System: 2.325 s]
  Range (min … max):   18.332 s … 18.511 s    10 runs

Benchmark 2: Before Optimization
  Time (mean ± σ):     27.027 s ±  0.062 s    [User: 24.668 s, System: 2.271 s]
  Range (min … max):   26.912 s … 27.105 s    10 runs

Summary
  After Optimization ran
    1.46 ± 0.01 times faster than Before Optimization
```

This PR introduces the concepts of a `Translator` and a `BatchedCollector`.

The `Translator` trait maps from hybrid RLE encoded values to an arbitrary set
of values. The HybridRLEDecoder will can then collect and call the translator
with batches of values. This way we minimize the amount of iterator polls and
we do not need to allocate any more on the heap except the output buffer. This
does however mean that the whole HybridRLEDecoder needs to be aware of the
`Translator` trait.

Furthermore, the `HybridRLEDecoder` can now itself buffer, instead of using the
`BufferedHybridRleDecoderIter` that was used before. Again, this allows minimal
memory consumption and prevents constant polling.

The `BatchedCollector` is essentially a wrapper for the `Pushable` trait that
automatically optimizes sequential pushes of valid and invalid values. It also
allows for efficient skipping of values.

Overall, this change significantly speeds up the Parquet reader and extensive
testing was done to ensure that no invalid data gets produced. But it is
difficult to test all edge-cases.

From here, we can start incorporating the `BatchedCollector` and `Translator`
traits in more places. In general, the `HybridRleDecoder` iterator
implementation should effectively never be used.