feat(python): Enable collection with gpu engine
Introduce option to collect a query using the cudf_polars gpu engine.
By default this falls back transparently to the default cpu engine if
the query cannot be executed. Configuration of specifics of the GPU
engine can be controlled by passing a GPUEngine object to the new
`engine` argument to `collect`.

The import of cudf_polars is currently quite expensive (will improve
in the coming months), so since we lazy-load the gpu engine, the first
query executed on gpu will pay this (one-time) cost.