perf: Parquet introduce MemReader to file buffer
This PR introduces three structures:

- `MemReader`: Abstraction over part of a Parquet file loaded into memory
- `MemReaderSlice`: A slice of a `MemReader`. This should should not be kept around outside the Parquet crate.
- `CowBuffer`: A Cow that abstracts between a `MemReaderSlice` and a `Vec<u8>`.

Following this crate, we can avoid copying the memory around in the Parquet crate. This also allows us to guarantee that the memory is in RAM and does not need to be loaded from disk anymore. Leading to less page faults.

Later it might be useful to make the reads a bit more granular so that we don't need to load unnecessary data pages.