perf: integer fast path Parquet dict encoding
This adds a fast path for the slowest part of parquet writing which is
(attemping) the dictionary encoding. Normally this involves a lot of hashing
which is quite slow. With this PR, it will first look at the minimum and
maximum values for integers. If these are sufficiently close together, we
create the dictionary using those values.