fix: Parquet nested values that span several pages
Fixes #18400.

This completely rewrites the logic that deal with decoding filtered nested
values. This was needed to accommodate the fact that any kind of row value
might span several data pages. This was not really accounted for in the
previous implementation and where it worked, it was mostly by coincidence.

This also made it possible to add many fast paths for the decoder that were
previously not there. Therefore, I think there is also some performance
improvements here, but I am not sure how much.

I also wrote two tests to make sure that it actually works.