fix(python): Change method op type hints to match dunder op type hints
Type hints for operators were refined in #13635 [0], but the refinements
were generally not propagated to the method equivalents (pow() being the
sole exception). This can result in equivalent expressions being treated
differently by type checkers, especially if stricter settings are used
that discourage/disallow the use of Any.

This commit changes the type hints for the method equivalents to match
those for the operators. This should ensure resulting types are
consistent regardless of whether one uses infix operators or their
method equivalents.

This also has the side benefit of making the documentation a bit more
consistent as the type hint more closely matches the stated type of the
argument in the docstring.

[0]: https://github.com/pola-rs/polars/pull/13635/commits/1c417e44cd7297d89a6931c50b3df565d3ee04e4