Simplify code example in Reading/Writing section
Simplify the dataframe used in the section 'Reading & Writing' of the getting started page and remove dead code from the respective source files. A smaller DF is better because the focus is on reading/writing, not on the data types supported.
It also creates a clearer distinction between the DF from this section and the DF of the section on expressions.