fix: Infer reshape dims when determining schema
This fixes the `reshape` function to properly infer dimensions in the schema.

I did this by adding a "zero-cost" `ReshapeDimension` enum that can either be
`Specified(u64)` or `Infer`. This also makes the reshaping behavior more
consistent as now all negative values are always assumed to be `Infer`.

During this work, I saw an unneeded allocation in `reshape_array` and removed
it.

Fixes #18879.