Download data in the Python script.
I was downloading the data in a separate hook but there is a test that just runs all of the Python scripts in src/python and apparently it's not trivial to get the data download to trigger just once, when the appropriate Python script test is running, so it's just easier to move the data download to the script that actually uses it. We also add a comment to the Rust script to direct users to the Python script for the data location.