fix(rust): Run join type coercion with correct schemas active
We previously ran type coercion for the left (respectively right) key
expressions with the schema of the output join node active. This is
incorrect since the keys may not even refer to column names in the
output (if the expressions do not refer to columns). Instead, the
left (right) key expressions must be coerced with the left (right)
inputs active.

With this done, we now no longer need to run type coercion on the
resulting Join node, and instead must just add it to the arena.

- Fixes #19597