docs: Assorted fixes to Rust API docs
I added two feature flags to make sure they're included in the docs and when I was adding a bit of detail to a docstring I noticed a link was formatted in the wrong way. Ended up using a regex to find all such links across the Rust API docs.