feat: Add `assert_err` and `assert_warn`.
This PR adds the `assert_eq` and `assert_warn` methods on `LazyFrame` and
`DataFrame`. These methods take a set of (possibly named) properties and verify
during execution that these properties hold for all values. If a property does
not hold, it will throw an error or a warning, respectively.

These methods are currently unstable, but I think they can be marked as stable
relatively quickly.