Fix subtraction with underflow on empty FixedSizeBinaryArray
When trying to deserialize an arrow ipc stream with a fixed size binary
array that contains no values, the code currently subtracts with
underflow. In the python package (which uses a release build), this is
silently ignored as debug assertions are turned off. Instead, you get a
panic about an out of bounds offset.

By checking for an underflow, everythning work fine.