refactor: Move hive partitioning outside of readers
This is a very early stage PR exploring whether it is possible to take the hive
partitioning out of the readers. This will also make the Hive partitioning
complete file type agnostic.

Currently, this PR does not do anything unless `POLARS_NEW_HIVE` is set to `1`.

Here is a rough todo list.

- [x] Row Indexes
- [x] Include File Paths
- [x] Projection Pushdown
- [ ] Slicing
- [ ] Hive Predicates
- [ ] Lazy Loading of Non-Hive data
- [ ] Allow Missing Columns
- [ ] Other file types besides Parquet
- [ ] New streaming engine sink