perf: Provide a fallback skip batch predicate for constant batches
This PR adds a fallback for skip batch predicate if we don't have a better
specialized implementation. Namely any expression that does not have a better
fallback now gets lowered to:

```
E -> all(col(A_min) == col(A_max) & col(A_nc) == 0 for A in LIVE(E)) & ~(E)
```

This basically means that if the predicate columns are constant for a batch, we
now are always accurately predict whether we can skip it.

Specifically, this makes pruning hive partitions much more consistent and
potent.

Fixes #21472.