refactor: Only use chrono_tz timezones in hypothesis testing
Rather than reading the valid timezones (and aliases) from the local
machine's zoneinfo, chrono_tz bakes the known timezones in at build
time. This can result in hypothesis testing producing timezones in
python that are unknown on the rust side.

To avoid this, filter out timezones that are unknown when sampling.