perf: Enable *Common Subplan Elimination* across plans in `collect_all`.
This is a long requested feature that allows for caching repeated computations
in a single `collect_all`. In the in-memory engine, this means less repeated
computation if LazyFrames perform partially the same computations. For the new
streaming engine, it is all combined into a single physical plan.

Fixes #5821.
Fixes #8369.
Fixes #13065.